+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Add reverse dependency tracking
+
 # 0.1.0.0
 
 - Initial release